Move filter functionality to filter method on APIFeatures.
APIFeatures is a Generic Class that takes in a type. This type will be the kind of data that this APIFeatures will be an instance of. For example, when working with the TourModel, it will be expecting the Tour interface to know what kind of data will be expected in the Query. It accepts two parameters, the Query (Expecting a resultType of a Tour array and a DocType of a Tour) and the query string (specified with the QueryArgs type).